+++ packages/core/src/index.ts
@@ -6,3 +6,4 @@ export * from "./installer.js";
 export * from "./packs.js";
 export * from "./new-skill.js";
 export * from "./runtime.js";
+export * from "./remote-registry.js";